Add BTPANHelper app
-Not sure what's going on with bluetooth tethering or why it
is in it's current state. It seems like it's not "really" supported.
The only real problem I found with it is that on the client side.
There is no method for the bt-pan to get an ip address and there
is not dns implementation unless another interface ( such as wlan0 )
has been used.

-I created this app to that will detect when a bluetooth connection
is made run dhcp on bt-pan, and run dnsmasq with google's dns
settings, allowing bluetooth tether to work.

-@Meticulus