stagefright: allow dynamic switching of h.264 decoder.
-Recent versions of the offical Youtube app are causing issues
with our proprietary OMX vendor blobs. However, it works fine with
the google software h.264 decoder. The solution seems simple, switch
to the software decoder. Well, the software decoder works with
the Youtube app but it doesn't do a good job with HD video. This
patch allows the switching for the hardware STE h.264 decoder and
the software Google h.264 decoder.

-Use property vu.co.meticulus.h264switch
--"false" for hardware STE decoder.
--"true" for software Google decoder.