recovery: fix blank screen (non - twrp)
-This allows cm's recovery to be seen but does not quite
fix it completely. Unless I am mistaken, The recovery menu
is not being drawn and can not be used although the header
with the words "cyanogen" and "Recovery" are seen. I suspect
it has something to do with the pixel format RGBA_8888 which
is reported by fb0  when it is probably RGB_565 that we really
need to use but cm's recovery has no support for...`